Console Application Assignment
Takes an input from the user, multiplies it by 50, then prints the result to the console.  Takes an input from the user, adds 25 to it, then prints the result to the console.  Takes an input from the user, divides it by 12.5, then prints the result to the console.  Takes an input from the user, checks if it is greater than 50, then prints the true/false result to the console.  Takes an input from the user, divides it by 7, then prints the remainder to the console.